requirements.txt
streamlit
pandas
openpyxl
transformers
torch
PyMuPDF
pytesseract
pillow
tqdm
